refactor: 서브 페이지 헤더(뒤로가기+제목) 공용 컴포넌트로 추출
- components/ui/SubPageHeader 추가
- TeamProfilePage, DepartmentMobile, CategoryDetailMobile에서 중복 구현하던
  뒤로가기 버튼 + 제목 헤더 마크업/스타일을 공용 컴포넌트로 교체
- 겸사겸사 Department 쪽에 남아있던 전역 font-family 중복 선언 제거
